requirements.txt
attrs==25.1.0
av==14.2.0
certifi==2025.1.31
cffi==1.17.1
ffmpeg-python==0.2.0
ffpyplayer==4.5.2
future==1.0.0
h11==0.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
mutagen==1.47.0
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
pycparser==2.22
pydub==0.25.1
pygame==2.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0